mm: /proc/sys/vm/stat_refresh to force vmstat update
Provide /proc/sys/vm/stat_refresh to force an immediate update of per-cpu
into global vmstats: useful to avoid a sleep(2) or whatever before
checking counts when testing.  Originally added to work around a bug which
left counts stranded indefinitely on a cpu going idle (an inaccuracy
magnified when small below-batch numbers represent "huge" amounts of
memory), but I believe that bug is now fixed: nonetheless, this is still a
useful knob.

Its schedule_on_each_cpu() is probably too expensive just to fold into
reading /proc/meminfo itself: give this mode 0600 to prevent abuse.  Allow
a write or a read to do the same: nothing to read, but "grep -h Shmem
/proc/sys/vm/stat_refresh /proc/meminfo" is convenient.  Oh, and since
global_page_state() itself is careful to disguise any underflow as 0, hack
in an "Invalid argument" and pr_warn() if a counter is negative after the
refresh - this helped to fix a misaccounting of NR_ISOLATED_FILE in my
migration code.

But on recent kernels, I find that NR_ALLOC_BATCH and NR_PAGES_SCANNED
often go negative some of the time.  I have not yet worked out why, but
have no evidence that it's actually harmful.  Punt for the moment by just
ignoring the anomaly on those.

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: Mel Gorman <mgorman@techsingularity.net>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Konstantin Khlebnikov <koct9i@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>