mm: thp: calculate the mapcount correctly for THP pages during WP faults
This will provide fully accuracy to the mapcount calculation in the write
protect faults, so page pinning will not get broken by false positive
copy-on-writes.

total_mapcount() isn't the right calculation needed in reuse_swap_page(),
so this introduces a page_trans_huge_mapcount() that is effectively the
full accurate return value for page_mapcount() if dealing with Transparent
Hugepages, however we only use the page_trans_huge_mapcount() during COW
faults where it strictly needed, due to its higher runtime cost.

This also provide at practical zero cost the total_mapcount information
which is needed to know if we can still relocate the page anon_vma to the
local vma.  If page_trans_huge_mapcount() returns 1 we can reuse the page
no matter if it's a pte or a pmd_trans_huge triggering the fault, but we
can only relocate the page anon_vma to the local vma->anon_vma if we're
sure it's only this "vma" mapping the whole THP physical range.

Kirill A.  Shutemov discovered the problem with moving the page anon_vma
to the local vma->anon_vma in a previous version of this patch and another
problem in the way page_move_anon_rmap() was called.

Mike Marciniszyn said:

: Our RDMA tests are seeing an issue with memory locking that bisects to
: commit 61f5d698cc97 ("mm: re-enable THP")
: 
: The test program registers two rather large MRs (512M) and RDMA
: writes data to a passive peer using the first and RDMA reads it back
: into the second MR and compares that data.  The sizes are chosen randomly
: between 0 and 1024 bytes.
: 
: The test will get through a few (<= 4 iterations) and then gets a
: compare error.
: 
: Tracing indicates the kernel logical addresses associated with the individual
: pages at registration ARE correct , the data in the "RDMA read response only"
: packets ARE correct.
: 
: The "corruption" occurs when the packet crosse two pages that are not physically
: contiguous.   The second page reads back as zero in the program.
: 
: It looks like the user VA at the point of the compare error no longer points to
: the same physical address as was registered.
: 
: This patch totally resolves the issue!

Link: http://lkml.kernel.org/r/1462547040-1737-2-git-send-email-aarcange@redhat.com
Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Reviewed-by: "Kirill A. Shutemov" <kirill@shutemov.name>
Tested-by: Alex Williamson <alex.williamson@redhat.com>
Tested-by: Mike Marciniszyn <mike.marciniszyn@intel.com>
Tested-by: Josh Collier <josh.d.collier@intel.com>
Cc: <stable@vger.kernel.org>	[4.5]
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>