mm: use phys_addr_t for reserve_bootmem_region() arguments
Since 92923ca3aacef63 ("mm: meminit: only set page reserved in the
memblock region") the reserved bit is set on reserved memblock regions. 
However start and end address are passed as unsigned long.  This is only
32bit on i386, so it can end up marking the wrong pages reserved for
ranges at 4GB and above.

This was observed on a 32bit Xen dom0 which was booted with initial memory
set to a value below 4G but allowing to balloon in memory (dom0_mem=1024M
for example).  This would define a reserved bootmem region for the
additional memory (for example on a 8GB system there was a reverved region
covering the 4GB-8GB range).  But since the addresses were passed on as
unsigned long, this was actually marking all pages from 0 to 4GB as
reserved.

Fixes: 92923ca3aacef63 ("mm: meminit: only set page reserved in the memblock region")
Link: http://lkml.kernel.org/r/1463491221-10573-1-git-send-email-stefan.bader@canonical.com
Signed-off-by: Stefan Bader <stefan.bader@canonical.com>
Cc: <stable@vger.kernel.org>	[4.2+]
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>